requirements.txt
connexion>=2.2
swagger-ui-bundle==0.0.5
pony>=0.7.10
psycopg2-binary>=2.8
pylint>=2.3
autopep8==1.4
flask_cors>=3.0.8
connexion[swagger-ui]
allure-pytest>=2.8.6
termcc>=2019.8.2
logbook>=1.5.2
pyjwt==1.7.1